% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_raw_data.R
\name{clean_ncc_sheet}
\alias{clean_ncc_sheet}
\title{clean_ncc_sheet}
\usage{
clean_ncc_sheet(path, sheet)
}
\arguments{
\item{path}{path to the raw excel file}

\item{sheet}{name of the sheet}
}
\value{
a tbl
}
\description{
clean_ncc_sheet
}
\examples{

}
